Remove redundant variables
Removed variables not needed and remade the remaining variables so they could be displayed on the screen.

The original project started in 1.4 and when you import a Scratch 1.4 program (.sb) in to Scratch 2 the variables are not visible on screen.  I think they're off the top left.
Only work around is to delete the variables and create them again. No need to change the code, just delete and create the variables fixes it.